+++ SBVUnitTest/SBVUnitTestBuildTime.hs
@@ -2,4 +2,4 @@
 module SBVUnitTestBuildTime (buildTime) where
 
 buildTime :: String
-buildTime = "Mon Nov 30 22:24:46 PST 2015"
+buildTime = "Sat Dec  5 16:41:14 PST 2015"